Fix: 자신의 코멘트만 수정/삭제할 수 있도록 백엔드에서도 처리 - 세션 id와 comment user id 가 같은지 확인 후 수정, 삭제 진행

+++ views/exhibition/exhibition.py
@@ -62,7 +62,7 @@ def edit_comment(id, comment_id):
     edited_content = request.form['edited_content']
     comment = db.session.get(Comment, comment_id)
 
-    if comment:
+    if comment and user_id == comment.user_id:
         comment.content = edited_content
         db.session.commit()
 
@@ -73,7 +73,7 @@ def delete_comment(id, comment_id):
     user_id = session.get('user_id')
     comment = db.session.get(Comment, comment_id)
 
-    if comment:        
+    if comment and user_id == comment.user_id:        
         db.session.delete(comment)
         db.session.commit()
